+++ scripts/utils.js
@@ -32,16 +32,33 @@ async function kill(child, signal = 'SIGINT') {
   } else {
     process.kill(-child.pid, signal);
   }
+
   try {
     await waitOnChild(child);
   } catch (err) {
-    // Should error if the error is not a child process error or it is a child
-    // process and either the platform is Windows or the signal matches.
-    const shouldError = err.name !== 'ChildProcessError' || (process.platform !== 'win32' && err.signal !== signal);
-    if (shouldError) {
-      throw err;
-    }
+    const signalNumber = getSignalNumber(signal);
+
+    // Ignore the error if it simply indicates process termination due to the signal we just sent.
+    // On Unix, a process may complete with exit code 128 + signal number to indicate
+    // that it was terminated by a signal. But we have the signal name.
+    const shouldIgnore =
+      err instanceof ChildProcessError &&
+      (process.platform === 'win32' || err.signal === signal || err.code === 128 + signalNumber);
+
+    if (!shouldIgnore) throw err;
   }
+}
+
+function getSignalNumber(signal) {
+  // We don't need any other signals for now, and probably never will, so no need to list them all.
+  // But if any other signals ends up being needed, look at `man 3 signal` for the complete list.
+  const SIGNAL_NUMBERS = {
+    SIGINT: 2,
+    SIGTERM: 15,
+  };
+
+  if (signal in SIGNAL_NUMBERS) return SIGNAL_NUMBERS[signal];
+  throw new TypeError(`Unknown signal in getSignalNumber: '${signal}'. Please add a case for that signal.`);
 }
 
 async function spawnNpx(args, opts) {
